Refactor + endpoint de comparativa en dinámica
